Fade Algorithm Rough
Rough fade algorithm successfully operational for address one.  Could theoretically be expanded to other addresses as well.  Works only if the difference in address value is greater than the fade time value (in tenths of a second) and only fades down.  A simple check could be implemented to add instead of subtract for a fade up.  Timing is less accurate the greater the time is as a remainder when determining dval results in a longer fade time than expected.  This could be resolved by implementing a "leap" every t/(s%t) frames that subtracts an additional 1.